// src/app/interfaces/fields-setup-maintenance.ts
import { IdDescription } from './id-description';

export class FieldsSetupMainInputHelp {
  a2zaccfieldsdto: FieldsSetupMainViewModel[];
  FieldsFlagList: IdDescription[];
}

export class FieldsSetupMainViewModel {
  Flag: number;
  Code: number;
  Description: String;
}

export class FieldsSetupDetailsCode {
  Flag: Number;
  Code: Number;
  Description: String;
}